Handle nonsensical zeros consistently.
Everything goes to zero when the prob does. Did not want to handle it in
the main module it complicates the logic and maybe this isn't quite
right for everything yet.

In some cases things should be going to infinity (or have generate an
error).. but I don't want to get into that just yet.