Improve probability handling.
The old format was a remnant of when I did a bad job of differentiating
between display values and computed values.  Now that I have that
squared away we should make the display nicer.

Includes letting us enter probabilities as a percentage.  This sounds
crazy, but it makes the inputs and outputs compatible.

The destring thing is still super ugly and surely non-idiomatic.  I want
to revist that at some point.

Also some whitespace fixes in the tests to make it easier to follow
along.